Add more robust error handling in image services
This commit introduces a new optional type, `ImageResponse`, to image-related services. This change allows us to return `null`, instead of throwing an error when an operation like delete, create, or get image fails. This new, more graceful error-handling approach leads to a better user experience and less risk of unhandled exceptions.

+++ src/lib/services/images/server-methods.ts
@@ -57,7 +57,7 @@ export const clearImage = async (id: string) => {
   if (!UPLOAD_DIR) throw new Error("Upload dir not found");
   try {
     const deletedImage = await deleteImage(id);
-    if (deletedImage.path) {
+    if (deletedImage && deletedImage.path) {
       await unlink(join(process.cwd(), UPLOAD_DIR, deletedImage.path));
     }
     return deletedImage;
